add feature checkLists

+++ src/main/java/com/example/QASystem/exceptions/CheckListNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.QASystem.exceptions;
 
 public class CheckListNotFoundException extends RuntimeException {
 
-    public CheckListNotFoundException(Integer id) {
+    public CheckListNotFoundException(Long id) {
         super("Чек-лист с id='" + id + "' отсутствует");
     }
 }
