+++ src/main/java/com/example/QASystem/exceptions/ProjectNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.QASystem.exceptions;
 
 public class ProjectNotFoundException extends RuntimeException {
 
-    public ProjectNotFoundException(Integer id) {
+    public ProjectNotFoundException(Long id) {
         super("Проект с id='" + id + "' отсутствует");
     }
 }
