+++ src/main/java/com/example/QASystem/repositories/TestCaseRepository.java
@@ -1,15 +1,14 @@
 package com.example.QASystem.repositories;
 
 
-import com.example.QASystem.model.Project;
 import com.example.QASystem.model.TestCase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TestCaseRepository extends JpaRepository<TestCase, Integer> {
+public interface TestCaseRepository extends JpaRepository<TestCase, Long> {
 
-    List<TestCase> getAllByProjectId(Integer id);
+    List<TestCase> getAllByProjectId(Long id);
 
 }
